Mycode/First/PythonProject/requirements.txt
acres==0.5.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
decorator==5.2.1
distro==1.9.0
easyocr==1.7.2
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.6
openai==1.90.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathlib==1.0.1
pillow==11.2.1
prompt_toolkit==3.0.51
prov==2.0.2
pure_eval==0.2.3
puremagic==1.29
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
tesseract==0.1.3
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
